Implemented the XNA Alien Game sample's ScreenManager.
Removed the Zune Input Manager and the Menu Game component to allow greater compatibility with the ScreenManager. While the code has been greatly cleaned up, touch support, the inability to select hidden tiles, the ability to scroll, and the game trying to prevent accidental clicking are all missing in this revision.